ref(commitlint/prettier): extract generic huskyrc rule
