+++ src/lib/rules/husky/apply-husky-configuration.ts
@@ -2,9 +2,9 @@ import { Rule, Tree } from '@angular-devkit/schematics';
 import { createHuskyRc } from './create-husky-rc';
 import { updateHuskyRc } from './update-husky-rc';
 
-export function applyHuskyConfiguration(): Rule {
+export function applyHuskyConfiguration(newHook: [string, string]): Rule {
   return (tree: Tree) =>
     !tree.exists('.huskyrc')
       ? createHuskyRc()
-      : updateHuskyRc(tree.read('.huskyrc')!);
+      : updateHuskyRc(tree.read('.huskyrc')!, newHook);
 }
