+++ src/lib/rules/husky/create-husky-rc.ts
