+++ src/lib/rules/husky/index.ts
