--- src/prettier/rules/husky/index.ts
@@ -1,1 +1,0 @@
-export * from './apply-husky-configuration';
